JAMES-2176 MailboxManager resolves users mailbox for SearchBackend
Before this commit SearchBackend had the responsibility to transform
MultimailboxesSearchQuery into a final list of mailboxes. This leaded to code
duplication and issue now than search backend can't rely anymore on user
property to answer the search won't read message not readable by the user.

Now search backend do not have this responsability to figure out which mailbox a
user can search, it just take a final list of mailbox and search inside it.

The job of transforming MultimailboxesSearchQuery into a final list of mailbox
is now done at the SearchBackend level.